requirements.txt
Django==1.7
Pillow==2.7.0
dj-database-url==0.3.0
dj-static==0.0.6
django-toolbelt==0.0.1
gunicorn==19.1.1
psycopg2==2.5.4
python-magic==0.4.6
requests==2.5.1
static3==0.5.1
whitenoise==1.0.6
